refactor: move ablytest under internal/ to make it unimportable externally
The ablytest package is test scaffolding for ably-go's own suite — in
particular its sandbox helpers (NewSandbox etc.) provision test apps against
the Ably backend, which is meant for developing and contributing to the SDK,
not as an API for consumers of the package.

Moving it to internal/ablytest makes Go's toolchain reject any import from
outside this module.

This is not a security boundary (the code can obviously be copied for a
start), but rather an unambiguous request. We don't want to actually
put any security barries in place, because we want SDK development to be
open to all, and there's no way to verify whether a given test app will
be used for SDK development or not. But at least we can make it
unambiguous and avoid any unintentional reliance
