+++ ably/message_updates_integration_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ablytest"
+	"github.com/ably/ably-go/internal/ablytest"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
