+++ ably/realtime_channel_delta_integration_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ablytest"
+	"github.com/ably/ably-go/internal/ablytest"
 	"github.com/stretchr/testify/assert"
 )
