+++ ably/realtime_channel_spec_integration_test.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ablytest"
+	"github.com/ably/ably-go/internal/ablytest"
 
 	"github.com/stretchr/testify/assert"
 )
