+++ ably/transitioner_integration_test.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ablytest"
+	"github.com/ably/ably-go/internal/ablytest"
 )
 
 type ConnTransitioner struct {
